Upload website trên hosting window + 1/2 thanh toan qua ngan luong

+++ TeduShop.Web/Assets/client/js/controllers/shoppingCart.js
@@ -85,6 +85,19 @@
             }
 
         });
+        $('input[name="paymentMethod"]').off('click').on('click', function () {
+            if ($(this).val() == 'NL') {
+                $('.boxContent').hide();
+                $('#nganluongContent').show();
+            }
+            else if ($(this).val() == 'ATM_ONLINE') {
+                $('.boxContent').hide();
+                $('#bankContent').show();
+            }
+            else {      
+                $('.boxContent').hide();
+            }
+         });
 
     },
     getLoginUser: function () {
@@ -110,7 +123,8 @@
             CustomerEmail : $('#txtEmail').val(),
             CustomerMobile : $('#txtPhone').val(),
             CustomerMessage : $('#txtMessage').val(),
-            PaymentMethod: "Thanh toán tiền mặt",
+            PaymentMethod: $('input[name="paymentMethod"]:checked').val(), 
+            BankCode: $('input[groupname="bankcode"]:checked').prop('id'),
             Status: false
         }
         $.ajax({
@@ -122,13 +136,18 @@
             },
             success: function (response) {
                 if (response.status) {
-                    console.log('create order ok');
-                    $('#divCheckout').hide();
-                    cart.deleteAll();
-                    setTimeout(function () {
-                        $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất.');
-                    }, 1500);
+                    if (response.urlCheckout != undefined && response.urlCheckout != '') {
+                        window.location.href = response.urlCheckout
+                    }
+                    else {
+                        console.log('create order ok');
+                        $('#divCheckout').hide();
+                        cart.deleteAll();
+                        setTimeout(function () {
+                            $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất.');
+                        }, 1500);
 
+                    }
                 }
             }
         });
